[CIR][CIRGen][Builtin][X86] Lower AVX masked load intrinsics (#1763)
For these intrinsics there only seems to be one function where the IR
emmited seems to diverge:

for `_mm_load_sbh` loads a single 16-bit bfloat (__bf16) value from
memory into the lowest element of a 128-bit bfloat vector (__m128bh),
leaving the remaining lanes unchanged or filled with a passthrough
value. It is implemented using a masked load with only the first lane
enabled.

[source for intrinsics with similar
behaviour](https://gist.github.com/leopck/86799fee6ceb9649d0ebe32c1c6e5b85)

In the CIR lowering of `_mm_load_sbh`, we are currently emitting the
mask of intrinsic (`llvm.masked.load`) operand as an explicit constant
vector:

``` llvm
<8 x i1> <true, false, false, false, false, false, false, false>
```
whereas OG lowers:
```llvm
<8 x i1> bitcast (<1 x i8> splat (i8 1) to <8 x i1>)
```
I believe both things are semantically equal so:

Is it acceptable for CIR and OG to diverge in this way for masked loads,
or should we aim for parity in how the mask is represented, even if that
reduces readability in CIR?