[CIR][LoweringPrepare] Wrap `cir.va_arg` lowered code in a `cir.scope`
The `cir.va_arg` lowering may introduce an if-else control flow.
This breaks the verifier for control flow operations, e.g., `cir.for`, where there can be only one block inside of the `cond` and `step` regions.
This PR wraps the lowered code in a `cir.scope` (only if it will lower to the if-else control flow).
Added a test for `va_arg` inside for loop condition, the basic pattern is: `for (; va_arg(...););`