[CIR] Use cir.get_vptr to initialize vptr members (#1784)
Previously, following classic codegen, we were using a bitcast to get
the address of the vptr member when initializing the vptr in a
constructor. This change updates the handling to use cir.get_vptr
instead.

This is functionally equivalent because the vptr is always at offset
zero for the CXXABIs we currently support, but using cir.get_vptr makes
what we are doing more explicit.