[CIR] Fix parsing of a generic TBAAAttr (#1846)
This PR fixes an issue when parsing `TBAAAttr` derived attributes.
The root cause of the issue is that the RTTI system is not capable of
downcasting `mlir::Attribute` to the `BaseType`, which is
`cir::TBAAAttr`, since each attribute has its unique TypeID. As such,
the default generated parser fails, because the `dyn_cast` to
`cir::TBAAAttr` fails (this happens for example with the attached test
case).
If anyone can think of a solution that wouldn't require explicitly
listing all the attributes I'm happy to modify it.