[CIR][CIRGen][Builtin][X86] Lower `insert*` intrinsics (#1829)
Only one thing to note related to the shuffles performed for these
intrinsics. We could tackle this in a different PR tho.

OG's builder creates unary shuffles by passing a poison value for the
second vector operand.

```cpp

/// Create a unary shuffle. The second vector operand of the IR instruction
/// is poison.
Value *CreateShuffleVector(Value *V, ArrayRef<int> Mask,
                             const Twine &Name = “”) {
    return CreateShuffleVector(V, PoisonValue::get(V->getType()), Mask, Name);
}

```

In CIR, we currently work around this by duplicating the single vector
argument as both operands:

```cpp
    cir::VecShuffleOp createVecShuffle(mlir::Location loc, mlir::Value vec1,
                                     llvm::ArrayRef<int64_t> mask) {
    // FIXME(cir): Support use cir.vec.shuffle with single vec
    // Workaround: pass Vec as both vec1 and vec2
    return createVecShuffle(loc, vec1, vec1, mask);
}

```

While this is just a minor nit and I don’t believe we provoke a semantic
difference, should we revisit the workaround stated in the comment for
parity with OG?