[CIR] Fix element type for lowering vtt.address_point (#1876)
While upstreaming vtt.address_point lowering, I noticed an important
difference in the LLVM IR generated between the CIR path and the OGCG
path. We were incorrectly using i8 as the element type for the GEP
generated when vtt.address_point is lowered with a value argument. This
patch fixes that.