[CIR][CUDA] FIx CUDA host compilation on kernel launch (#1906)
This PR implements some missing blocks that allow us to effectively
allow us to launch kernels from the host. All of the tests stated in
this
[commit](https://github.com/llvm/clangir/commit/69f2099870067dcacbc6fecb10c77f22e0c5b2d9)
are now resolved.

I spent half a day figuring the following:

I tried experiementing performing host compilation(`-fcuda-is-device`)
with target triple: `nvptx64-nvidia-cuda` but was getting a module
verification error that, to keep it simple looked like: `error:
'cir.call' op calling convention mismatch: expected ptx_kernel, but
provided c`.

I thought that was expected given that we're essentially using the
device to compile on the host, which doesn't make a lot of sense. until
I tried to replicate the same in OG and didn't really run into any
problem in that regard. Are the calling conventions enforced in CIR much
more strict as compared to OG? Or is that simply a bug from OG?