[Docs] Reapply #1914 (#1916)
Seems like we're using github-pages gem and it doesn't include the theme
for just-the-docs, switching to remote_theme.

From
https://github.com/just-the-docs/just-the-docs/issues/1273#issuecomment-1595854665
> By default, when you create a GitHub repository and then enable GitHub
Pages, it is deployed with a Ruby gem called github-pages. Among other
things, this forces a certain version of Jekyll (version 3.9.3) and does
not run bundle install; it only uses a preselected set of gems. Since
just-the-docs is not one of them, you then get the error you
encountered.

and
https://talk.jekyllrb.com/t/yet-another-github-pages-build-error-thread/7241/3
> There is a small subset of 3rd party themes allowed by GitHub when
using the theme config. This is done for security reasons as installing
themes this way use Ruby gems and GH only allows ones that pass their
requirements.
Installing a theme via the remote_theme configuration (if available)
doesn’t have this limitation. Nor does using a GitHub action workflow to
build/deploy your Jekyll based site on GitHub Pages.

Hahha i'm praying it passes the CI this time... I don't know of another
way to go about this if this fails, since I can't reproduce the error
locally and do not have push permission for gh-pages branch