[CIR] Backport auto var constant init fix
This backports a problem I noticed while upstreaming constant initialization
of automatic aggregate variables. Even though the aggregate is being
initialized, we weren't setting the `init` attribute.

There is one odd case where the constant being used to initialize a
class is an undef. OGCG does the same thing. In this patch I'm not
attempting to detect that case, and I'm not sure if we should.