[CIR] Backport `TargetAddressSpaceAttr` and Support both existing Lang AS and target AS attributes in pointer types and Global Ops (#1986)
Bringing the changes we performed at:
https://github.com/llvm/llvm-project/pull/161028 down to the incubator
isn't as straight-forward as I though. Therefore — this PR might be a
bit long but hopefully not too complicated, bare with me :)

The main purpose of this is to bring the recent upstreamed
`TargetAddressSpaceAttr` and couple it with `PointerType` and
`GlobalOp`. The main challenge is that this collides with the already
implemented infrastructure related to AS that revolves around our
unified enum approach (That handles numeric and lang AS). My main
rationale here is that we let our new attribute to coexist with the
already existing `cir::AddressSpaceAttr` (renamed now). so that we don't
break any tests related to offload-type languages.

Considering the above what I'm essentially doing is:

1. Letting `TargetAddressSpaceAttr` handle numeric target AS as it is
done upstream.
2. Rename our previous `AddressSpaceAttr` to `LangAdddressSpaceAttr`.
3. Implement `MemorySpaceAttrInterface`for lang and target AS.
4. The current implementation of `LangAddressSpaceAttr`(which handles an
enum) will handle language/clang specific AS (CUDA, OpenCL, etc..) it
will no longer handle target as it used to!
5. `PointerType` and globalOps will hold a `MemorySpaceAttrInterface`.
Similar to the upstream ptr type in the ptr dialect.
6. Adjusted the assembly format for all tests containing any AS
attributes to conform with our new format:
`lang_address_space(n)`/`clang_address_space(x)`.
7. For the relevant Ops: Default address spaces will now be represented
as an empty `MemorySpaceAttrInterface` attribute.
8. Remove any comments that make reference to our old approach regarding
AS handling.
9. Updated description where relevant.
10. Bring any extra upstream changes that deviate with what we have
here.