[CIR][CIRGen] Fix missing annotation attributes on function decls
Currently under CIRGen, if a function declared with `__attribute__((annotate(...)))`
is encountered but no definition of this function is encountered, its annotation is
not propagated into the global annotations table.

This commit fixes this by ensuring that if we detect an `AnnotateAttr` on the
`FunctionDecl` at `cir::FuncOp` creation time, it is always added to the deferred
annotations map, aligning CIRGen behavior with clang codegen.

A new test has been added based on `clang/test/CodeGen/annotations-declaration.c`
but extended to check for propagation of annotations on function decls through
CIR and into LLVM.