[CIR][ThroughMLIR] Lower pointers to memref<?xType>
This changes the default type conversion of cir.ptr to memref<?xType>

To make the intermediate legalization work, this also changes the
lowering of AllocaOp and GlobalOp to generate memref<1xType> followed by
a CastOp to memref<?xType>. CastOp is later eliminated in
Load/Store-lowering. Load/Store-lowering's findBaseAndIndices was then
adjusted to generate op with zero indices in case a non-indexed op is
performed.

Add a testcase ptr-arg.cir, with example code similar to one in #1729

I also had to touch the loop lowering to SCF, since the alloca-change
affected the 'canonical-loop recognition'.

Fixes #1729