[CIR] Add ABI lowering pass (#1471)
This PR attempts to add a new pass `cir-abi-lowering` to the CIR
dialect. This pass runs before the CallConvLowering pass, and it expands
all ABI-dependent types and operations inside a function to their
ABI-independent equivalences according to the ABI specification.

The patch also moves the lowering code of the following types and
operations from the LLVM lowering conversion to the new pass:
  - The pointer-to-data-member type `cir.data_member`;
  - The pointer-to-member-function type `cir.method`;
  - All operations working on operands of the above types.